fix: use uint64 for series index

+++ pkg/backends/influxdb/model/unmarshal.go
@@ -145,7 +145,7 @@ func UnmarshalTimeseriesReader(reader io.Reader, trq *timeseries.TimeRangeQuery)
 			sh.FieldsList = make([]timeseries.FieldDefinition, fdl)
 			var fdi int
 			for ci, cn := range wfd.Results[i].SeriesList[j].Columns {
-				index := uint32(ci) // #nosec G115 -- ci is positive
+				index := uint64(ci) // #nosec G115 -- ci is positive
 				if cn == "time" || cn == "_time" {
 					timeFound = true
 					sh.TimestampIndex = index
@@ -233,7 +233,7 @@ func tryParseTimestamp(v any) int64 {
 	return -1
 }
 
-func pointFromValues(v []interface{}, tsIndex uint32) (dataset.Point,
+func pointFromValues(v []interface{}, tsIndex uint64) (dataset.Point,
 	[]timeseries.FieldDataType, error) {
 	p := dataset.Point{}
 	ns := tryParseTimestamp(v[tsIndex])
